requirements.txt
Flask[async]>=2.0
gunicorn>=20.0
google-generativeai>=0.5.0
